Implemented: Changes based testrun
Does not work yet, as I don't get phpunit to execute a bunch of provided unit tests. It only seems to execute the last one given :(
If this does not work an apply would be needed, but I don't really like that.

Furthermore there is a problem with source files, which do not have a corresponding test file yet. Those should be filtered out as well.